+++ src/components/form/ImportModelTemplateForm.vue
@@ -99,9 +99,8 @@ async function onSubmit() {
       });
 
       return router.push({
-        name: 'modelizer',
+        name: 'Draw',
         params: {
-          viewType: 'draw',
           projectName: props.projectName,
         },
         query: { path: `${props.template.plugin}/${modelName.value}` },
