+++ src/router/routes.js
@@ -1,5 +1,7 @@
-import ModelizerRoute from 'src/router/routes/ModelizerRoute';
 import ModelsRoute from 'src/router/routes/ModelsRoute';
+import ModelizerDrawLayout from 'src/layouts/ModelizerDrawLayout.vue';
+import ModelizerTextLayout from 'src/layouts/ModelizerTextLayout.vue';
+import { getProjectById } from 'src/composables/Project';
 
 const routes = [
   {
@@ -8,7 +10,6 @@ const routes = [
     children: [
       { path: '', name: 'Home', component: () => import('pages/HomePage.vue') },
       { path: '/about', name: 'About', component: () => import('pages/AboutPage.vue') },
-      ModelizerRoute,
       {
         path: '/:projectName/diagrams',
         name: 'Diagrams',
@@ -17,6 +18,30 @@ const routes = [
     ],
   },
   ModelsRoute,
+  {
+    path: '/projects/:projectName',
+    beforeEnter: (to, _from, next) => {
+      const project = getProjectById(to.params.projectName);
+
+      if (!project) {
+        next({ name: 'Error' });
+      } else {
+        next();
+      }
+    },
+    children: [
+      {
+        path: 'modelizer/draw',
+        name: 'Draw',
+        component: ModelizerDrawLayout,
+      },
+      {
+        path: 'modelizer/text',
+        name: 'Text',
+        component: ModelizerTextLayout,
+      },
+    ],
+  },
   // Always leave this as last one,
   // but you can also remove it
   {
